Import stylable's createGenerator directly from dist Stylable now exports utils that use Node's fs (given no alternative fs) and path, and requires ts users to have @types/node for validation to pass (when importing from main entry point).
As wix-react-tools only uses the generator for tests, and doesn't use any Node.js APIs, we now import it directly to avoid ts encountering .d.ts files that import fs/path. Saving us of adding @types/node when we dont actually use any node APIs.

This calls for a discussion of what we see fit for the main entry point of stylable to export.